refactor!: Rename file `mfa.py` to `mfa_config.py`
